configs: Enable pstore and ramoops as built-ins
The pstore and ramoops modules together allow kernel crash logs to be
preserved across a reboot. They have beeb configued as built-ins
(rather than as modules) because the crash dump facility becomes
available much earlier in the boot sequence, and a lot of new-kernel
crashes happen early.

Note that if systemd-pstore is enabled then the dmesg captures will
be moved to /var/lib/systemd/pstore/ after the reboot.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>